MA-5952 Add NTFS support in Lollipop
When plug in ntfs u disk,ntfsfix will fail.it can't call vfprintf and fprintf in logging.c
The root cause is that it define off_t as "long long" rather than "long int"

+++ config.h
@@ -361,7 +361,7 @@
 /* Define to `long int' if <sys/types.h> does not define. */
 #ifndef _OFF_T_DEFINED_
 #define _OFF_T_DEFINED_
-typedef long long off_t;
+typedef long int off_t;
 #endif
 
 /* Define to `unsigned int' if <sys/types.h> does not define. */
